Properly handle lemma of the underscore character.
The underscore (_) is normally lemmatized as an underscore. In CoNLL-X,
underscores are used for absent annotations. So, in these cases the
EditTreeEncoder would consider the lemma to be absent and fail with an
error.

This change uses `_` as a lemma if the form (which cannot be absent) is
the `_` character.