find_root_candidate: check that candidate attaches to ROOT token
Before this change find_root_candidate looked for an encoding that has
the root label and can be decoded.

Unfortunately, the Lassy treebank has some root edges that do not attach
to ROOT. find_root_candidate could find a root token that is not
attached to ROOT. To make things worse, in the case of a cycle between
such a token and another token, the cycle cannot be resolved.

This change adds an additional requirement to candidates, namely that
they should attach the token to ROOT after decoding.