break_cycles: assert that we are not in an infinite loop
The break_cycles function operates on the assumption that the given root
index does not participate in the cycle. If the root index is in the
cycle, the method could enter an infinite loop. Panic in such a loop by
asserting that we are making progress.

(cherry picked from commit 6cea5302c8a320e21ff517d9732986438ffd21b0)